+++ OpenEXR/IlmImfTest/testDeepScanLineMultipleRead.cpp
@@ -161,14 +161,14 @@ static void read_file(const char * filename)
         for(int i=0;i<width;i++)
         {
             
-            if( samplecounts[i]!= row)
+            if( samplecounts[i]!= static_cast<unsigned int>(row))
             {
               cout << i << ", " << row << " error, sample counts hould be "
               << row  << ", is " << samplecounts[i]
               << endl << flush;
             }
             
-            assert (samplecounts[i]== row);
+            assert (samplecounts[i]== static_cast<unsigned int>(row));
             
             total_samples+=samplecounts[i];
         }
